refactor(cli): remove production code only tests kept alive
With the speculative tests gone, golangci-lint's unused check exposed
production code that had no caller outside those tests: 56 functions, types
and constants deleted. Functions that only one platform's code calls moved
beside that caller (lifecycle guard helpers to darwin, wake-check schema
helpers to the non-unix build, the reload-transport response to linux) so the
unused check passes on every GOOS. No behavior change; build, vet and the
pinned linter are clean for linux, darwin and windows.
