+++ internal/cli/coop_wake_reclaim_unix.go
@@ -376,88 +376,6 @@ func coopWakeTTYDisplay(inspection wakeLockInspection) string {
 	return tty
 }
 
-// resolveMissingWakeLockAfterTermination handles the irreversible half of a
-// takeover: the exact old helper's lock disappeared during termination, but
-// the process path could not prove that the helper exited. With no lock left to
-// preserve, proceeding is safer than returning a false retry; warn that the old
-// helper may still emit duplicate notifications.
-func resolveMissingWakeLockAfterTermination(
-	inspection wakeLockInspection,
-	terminationErr error,
-) (resolved bool, retErr error) {
-	defer func() { retErr = withWakeDiagnostic(retErr, inspection.Root, inspection.Agent) }()
-	agentDir, err := openExistingCoopWakeAgentDir(inspection.Root, inspection.Agent)
-	if err != nil {
-		return false, err
-	}
-	if agentDir == nil {
-		return resolveMissingWakeLockAfterTerminationFromInspection(
-			inspection,
-			wakeLockInspection{},
-			terminationErr,
-		)
-	}
-	defer func() { _ = agentDir.Close() }()
-	return resolveMissingWakeLockAfterTerminationInDir(agentDir, inspection, terminationErr)
-}
-
-func resolveMissingWakeLockAfterTerminationInDir(
-	agentDir *wakeAgentDir,
-	inspection wakeLockInspection,
-	terminationErr error,
-) (resolved bool, retErr error) {
-	defer func() { retErr = withWakeDiagnostic(retErr, inspection.Root, inspection.Agent) }()
-	var current wakeLockInspection
-	if err := agentDir.withFD(func(dirfd int) error {
-		if err := validateWakeStateAgentDirAt(dirfd, agentDir); err != nil {
-			return err
-		}
-		current = inspectWakeLockAt(dirfd, agentDir, inspection.Root, inspection.Agent)
-		return validateWakeStateAgentDirAt(dirfd, agentDir)
-	}); err != nil {
-		relation, relationErr := retainedWakeAgentDirRelation(agentDir)
-		if relationErr != nil {
-			return false, errors.Join(err, relationErr)
-		}
-		switch relation {
-		case wakeAgentDirDetached:
-			return false, nil
-		case wakeAgentDirCanonical:
-			return false, err
-		case wakeAgentDirInconclusive:
-			return false, errors.Join(err, fmt.Errorf("wake agent directory relation is inconclusive after termination"))
-		default:
-			return false, errors.Join(err, fmt.Errorf("unknown wake agent directory relation %d", relation))
-		}
-	}
-	return resolveMissingWakeLockAfterTerminationFromInspection(inspection, current, terminationErr)
-}
-
-func resolveMissingWakeLockAfterTerminationFromInspection(
-	inspection wakeLockInspection,
-	current wakeLockInspection,
-	terminationErr error,
-) (resolved bool, retErr error) {
-	defer func() { retErr = withWakeDiagnostic(retErr, inspection.Root, inspection.Agent) }()
-	if current.Exists && !sameWakeLockGeneration(inspection, current) {
-		return false, nil
-	}
-	if current.Exists {
-		return false, terminationErr
-	}
-	// This PID-based inspection only decides whether to print the warning. Both
-	// outcomes proceed identically, so PID reuse cannot authorize an action.
-	if inspectWakeIdentity(inspection) != wakeIdentityGoneOrDifferent {
-		_ = writeStderr(
-			"warning: superseded wake helper for %s (pid %d on %s) after its lock disappeared during termination without a confirmed exit; fresh wake is starting, but duplicate notifications may continue until the old helper exits; stop that helper if duplicates persist\n",
-			inspection.Agent,
-			inspection.PID,
-			coopWakeTTYDisplay(inspection),
-		)
-	}
-	return true, nil
-}
-
 func coopWakeRemedy(inspection wakeLockInspection, state, command string) string {
 	inspect := newWakeRemedy("amq", "doctor", "--root", inspection.Root, "--ops").String()
 	return fmt.Sprintf("wake for %s is blocking startup and cannot notify you.\n  lock: %s\n  state: %s\n\nRe-run with -y to clear it and start a fresh wake:\n  %s\n\nTo inspect first:\n  %s", inspection.Agent, inspection.LockPath, state, command, inspect)
