+++ internal/cli/doctor_ops_fix_other.go
@@ -2,6 +2,8 @@
 
 package cli
 
+import "fmt"
+
 func fixStaleWakeLockForDoctor(
 	root string,
 	agent string,
@@ -27,4 +29,25 @@ func fixStaleWakeLockForDoctor(
 		lock.Removed = true
 		return nil
 	})
+}
+
+func validateWakeLockStaleRemoval(inspection wakeLockInspection) error {
+	if _, err := readWakeStateSelectionForInspection(
+		inspection.Root,
+		inspection.Agent,
+		inspection,
+	); err != nil {
+		return err
+	}
+	if wakeLockHasOwnerMarkers(inspection) {
+		return fmt.Errorf("owner-bound wake claims require %s", wakeRecoverOwnerCommand(inspection.Root, inspection.Agent))
+	}
+	if err := validateWakeLockRepairable(inspection); err == nil {
+		return nil
+	} else if inspection.Status != wakeLockStale {
+		return err
+	}
+	// Identity mismatches reach stale only when the tri-state classifier has
+	// affirmative proof that the recorded generation is gone or different.
+	return nil
 }
