+++ internal/cli/wake_lifecycle_guard_darwin.go
@@ -2,8 +2,29 @@
 
 package cli
 
-import "golang.org/x/sys/unix"
+import (
+	"errors"
+
+	"golang.org/x/sys/unix"
+)
 
 func withExistingWakeLifecycleGuardNoWaitInDir(agentDir *wakeAgentDir, fn func(int) error) error {
 	return withExistingWakeLifecycleGuardModeInDir(agentDir, unix.LOCK_EX|unix.LOCK_NB, fn)
+}
+
+func withExistingWakeLifecycleGuardModeInDir(
+	agentDir *wakeAgentDir,
+	lockMode int,
+	fn func(int) error,
+) error {
+	return withWakeLifecycleGuardLeaseModeAndTimeoutInDir(
+		agentDir,
+		true,
+		lockMode,
+		wakeLifecycleGuardRetryTimeout,
+		func(dirfd int, lease *wakeLifecycleGuardLease) (retErr error) {
+			defer func() { retErr = errors.Join(retErr, lease.release()) }()
+			return fn(dirfd)
+		},
+	)
 }
