+++ internal/cli/wake_mutation_scope_unix.go
@@ -293,23 +293,6 @@ const (
 	wakeAgentDirDetached
 )
 
-// retainedWakeAgentDirRelation distinguishes a proven detached retained
-// directory from a namespace lookup that failed for an unknown reason. It
-// reacquires agentDir.withFD, so call it only when no withFD callback is active;
-// use retainedWakeAgentDirRelationAt inside one.
-func retainedWakeAgentDirRelation(agentDir *wakeAgentDir) (wakeAgentDirRelation, error) {
-	if agentDir == nil {
-		return wakeAgentDirInconclusive, fmt.Errorf("wake agent directory capability is missing")
-	}
-	var relation wakeAgentDirRelation
-	err := agentDir.withFD(func(dirfd int) error {
-		var err error
-		relation, err = retainedWakeAgentDirRelationAt(agentDir, dirfd)
-		return err
-	})
-	return relation, err
-}
-
 // retainedWakeAgentDirRelationAt inspects an already-held directory descriptor
 // without acquiring agentDir.mu. Use it from an active withFD callback.
 func retainedWakeAgentDirRelationAt(
