+++ internal/cli/wake_restart_stage_linux.go
@@ -34,6 +34,4 @@ func wakeRestartStageExistsPlatform(wakeRestartRecord) (bool, error) {
 	return false, nil
 }
 
-func wakeRestartStageUsesStableStatePlatform(string) (bool, error) { return false, nil }
-
 func reclaimWakeRestartRunningImagePlatform(wakeLock) error { return nil }
