+++ internal/cli/wake_terminate_legacy_test_linux.go
@@ -12,10 +12,6 @@ var signalWakeProcess = func(int, os.Signal) error {
 	return errors.New("bare-PID wake signaling is unavailable on Linux")
 }
 
-func terminateWakeProcess(wakeLockInspection) error {
-	return errors.New("bare-PID wake termination is unavailable on Linux")
-}
-
 func requireBarePIDWakeTermination(t *testing.T) {
 	t.Helper()
 	t.Skip("bare-PID termination is Darwin-only; Linux pidfd behavior has platform tests")
